Simplify indexing process
It's unscable to index at startup along with making ~1.2k HTTP calls to
get stats plus possibly another ~1.2k HTTP calls to scrape HTML content.
The plan moving forward is to rely on the build process to generate JSON
file(s) that include all pertinent information so that all that has to
happen at startup is the app reads these files and performs simple
indexing.

In the future these files shouldn't be bundled with the WAR but hosted
on S3 thereby making it possible to freshen the data without having to
rebuild the app. Ideally we'll have a REST endpoint that tells the app
to reindex using the file from S3.