[WEBSITE-243] - Add mapping and ES version
This adds additional information to the generated plugin data to help
prevent ingestion of data that cannot be handled by the running
application.

The typical scenario will be adding or modifying
properties to the ES mapping, such as changing a scalar field to a
nested property. The currently running application expects a scalar
field and will bomb as a result.

The next scenario is updating the version of ES. While minor updates are
unlikely to cause problems this adds an additional check help protect
the API.