[WEBSITE-284] - Properly handle wiki failures
The problem is the LoadingCache doesn't allow null values to be
associated with a key. In this case the url (key) cannot be mapped to
null content. The fix is ultimately returning null for the result of
getWikiContent in the catch.

Improvements include removing redundant code for following redirects.
HttpClients will automatically follow unless told not to.